split out types.go file
